Removed the word cart, again
I don't like that word, it is very negative. I wish AI would not use hateful words like that. Earlier it tried to get me to call someone a mutt, I was shocked. Luckily the filters were up and we caught it. I guess there is more work to do on the filters. 

This one is weird because it's technically a term but so were many other words in the past, well 6-7 homies. Like young Reezy be sayin in the BD, "IT BE LIKE THAT SOMETIMES"